feat(D-16): #11-wasm-vm — WebAssembly JS API on VM (#246)
Layered WebAssembly JS API on the elidex-js VM atop F1/F2/F3 engine-bridge
surfaces (`elidex-wasm-runtime`):

- `WebAssembly` namespace + `validate` / `compile` / `instantiate` static methods
- 5 constructors: `Module`, `Instance`, `Memory`, `Table`, `Global`
- 3 error subclasses: `CompileError`, `LinkError`, `RuntimeError`
- Exported function exotic (`ObjectKind::WasmExportedFunction`) with dispatch
  arm + GC trace
- Frozen `Instance.exports` namespace with wrapper-identity-stable cache
- `Memory.buffer` live-view aliasing (DR-11) via `wasm_backed_buffers`
  reverse-lookup + ArrayBuffer detach SoT encapsulation

Original 6-commit set (Stages 1-5 + 2 fix-delta commits) + 11 R-loop commits
(9 Copilot rounds 1-9 + .bak housekeeping + clippy CI fix).
44 wasm tests + 10299/10299 workspace nextest. Both wasm-runtime + wasm-vm
trip-wires wired into `mise run ci` via the `trip-wires` task.

Pre-push gate (5 stages): cargo fmt → mise run ci → /code-review (10 fix +
2 defer) → /review (verdict: ship) → /elidex-review (5-agent: F0 CRIT-class
revert + 19 citation corrections + 3 cleanups + 3 defers).  F0 catch:
/code-review Stage 3.5 Fix #11 inverted WASM JS API §5.5 Global-setter step
order; webref-verified spec mandates step 5 mut check BEFORE step 6
ToWebAssemblyValue.  Reverted in c2c43f07.

Copilot R-loop convergence (9 rounds, TERMINAL on R8+R9 zero-CRIT-zero-IMP):
- R1: 3 IMP (WebIDL §3.2.17 dict→DATA enumerability) + 5 doc-MIN
- R2: 3 doc-MIN (coupling-invariant `HashMap::get` notation)
- R3: 1 IMP test strict + 1 FP-docfix (EnforceRange clarification)
- R4: 1 MIN byte_io `u32::try_from` guard
- R5: 1 IMP structural — `array_buffer_view_bytes` signature `u32`→`usize`
  for > 4 GiB ArrayBuffer slice correctness
- R6: 1 doc-MIN + 1 FP-cite (no-engine cfg-gap codebase-wide condition)
- R7: 1 IMP + audit-found 1 IMP (Table.set/grow OOB → RangeError per
  WASM JS API §5.4 step 9) + 1 spec-MIN clone label
- R8: 3 doc-MIN (WasmError vs WasmErrorKind discipline sweep)
- R9: 2 doc-MIN (Global setter doc + `setup_bound_vm` rename)

Plus 2c8fa9f2 clippy fix (unused `WasmErrorKind` import — intra-doc-link
references don't count as use).

Defer ledger:

CLOSES `#11-wasm-vm` (−1).

NEW (+9): 5 from plan §7.1 (`#11-wasm-memory-resizable-buffer` /
`#11-wasm-threads` / `#11-wasm-exception-handling` / `#11-wasm-gc-proposal` /
`#11-wasm-esm-integration`) + 1 plan §0.5 E5
(`#11-wasm-memory-with-data-mut-encapsulation`) + 1 /elidex-review F23+F26
subsumption (`#11-wasm-extern-v128-result-marshalling`) + 1 /code-review #10
defer (`#11-vm-gc-enabled-raii-guard`) + 1 user-direction priority NEXT-UP
(`#11-vm-native-constructor-only-flag`) + 1 R7 audit-found
(`#11-wasm-memory-grow-return-minus-one` — Memory.grow should return -1 on
failure per WASM JS API §5.3, not throw RuntimeError).

Cap state: **47/35 (12 over)** per `feedback_cap-vs-completeness`. 5 of 10
NEW are spec-fork stage-gated; 2 are cross-cutting VM-wide refactors; rest
are scoped follow-ups.

Landing memo: `m4-12-pr-d16-wasm-vm-landing.md`.

+++ crates/script/elidex-js/src/vm/host/canvas/image_data.rs
@@ -512,7 +512,7 @@ fn read_typed_array_bytes(vm: &VmInner, id: ObjectId) -> Option<Vec<u8>> {
     Some(array_buffer::array_buffer_view_bytes(
         vm,
         buffer_id,
-        byte_offset,
-        byte_length,
+        byte_offset as usize,
+        byte_length as usize,
     ))
 }
