+++ crates/script/elidex-js/src/vm/host/request_response/body_init.rs
@@ -94,8 +94,8 @@ pub(in crate::vm::host) fn extract_body_bytes(
                     super::super::array_buffer::array_buffer_view_bytes(
                         ctx.vm,
                         buffer_id,
-                        byte_offset,
-                        byte_length,
+                        byte_offset as usize,
+                        byte_length as usize,
                     ),
                     None,
                 )))
